IFs and FORs In-Class Activity
Program that reads in an integer number from a user. If the number is divisible by 3, print out "Fizz". If the number is divisible by 5, print out "Buzz". If the number is divisible by 3 and 5, print out "FizzBuzz". Otherwise, print out the original number.

Program that prints out the first 100 positive integers (e.g. 1, 2, 3, 4, 5, ...)

Program that prints out the sum of the first 100 positive integers (e.g. 1 + 2 + 3 + 4 + ...)